дублювання карточок
