+++ src/index.js
@@ -1,4 +1,5 @@
 import './sass/index.scss';
+import './js/componentsJs/setLocalStorage'
 import { renderModalFilm } from './js/templates/renderModalFilm';
 import { getApiList } from './js/componentsJs/generateCardList';
 import callModal from './js/componentsJs/modalFooter';
